// src/Crease.WebUI/client-app/src/api/apiClient.ts
export interface IBankCardDto {
  id?: string | undefined;
  name?: string | undefined;
  bank?: Bank | undefined;
  statementType?: StatementType | undefined;
  transactionDateType?: TransactionDateType | undefined;
}

export abstract class ValueObject implements IValueObject {
  constructor(data?: IValueObject) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (<any>this)[property] = (<any>data)[property];
      }
    }
  }

  init(_data?: any) {}

  static fromJS(data: any): ValueObject {
    data = typeof data === 'object' ? data : {};
    throw new Error("The abstract class 'ValueObject' cannot be instantiated.");
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    return data;
  }
}

export interface IValueObject {}

export class Bank extends ValueObject implements IBank {
  name?: string | undefined;

  constructor(data?: IBank) {
    super(data);
  }

  init(_data?: any) {
    super.init(_data);
    if (_data) {
      this.name = _data['name'];
    }
  }

  static fromJS(data: any): Bank {
    data = typeof data === 'object' ? data : {};
    let result = new Bank();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['name'] = this.name;
    super.toJSON(data);
    return data;
  }
}

export interface IBank extends IValueObject {
  name?: string | undefined;
}

export class StatementType extends ValueObject implements IStatementType {
  value?: string | undefined;

  constructor(data?: IStatementType) {
    super(data);
  }

  init(_data?: any) {
    super.init(_data);
    if (_data) {
      this.value = _data['value'];
    }
  }

  static fromJS(data: any): StatementType {
    data = typeof data === 'object' ? data : {};
    let result = new StatementType();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['value'] = this.value;
    super.toJSON(data);
    return data;
  }
}

export interface IStatementType extends IValueObject {
  value?: string | undefined;
}

export class TransactionDateType
  extends ValueObject
  implements ITransactionDateType
{
  value?: string | undefined;

  constructor(data?: ITransactionDateType) {
    super(data);
  }

  init(_data?: any) {
    super.init(_data);
    if (_data) {
      this.value = _data['value'];
    }
  }

  static fromJS(data: any): TransactionDateType {
    data = typeof data === 'object' ? data : {};
    let result = new TransactionDateType();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['value'] = this.value;
    super.toJSON(data);
    return data;
  }
}

export interface ITransactionDateType extends IValueObject {
  value?: string | undefined;
}
// ...
export interface ICardDto {
  id?: string | undefined;
  name?: string | undefined;
  bankCardId?: string | undefined;
}
// ...
export class CreateCardCommand implements ICreateCardCommand {
  bankCardId?: string | undefined;
  name?: string | undefined;
  cardNumber?: string | undefined;
  startDate?: Date;

  constructor(data?: ICreateCardCommand) {
    if (data) {
      for (var property in data) {
        if (data.hasOwnProperty(property))
          (<any>this)[property] = (<any>data)[property];
      }
    }
  }

  init(_data?: any) {
    if (_data) {
      this.bankCardId = _data['bankCardId'];
      this.name = _data['name'];
      this.cardNumber = _data['cardNumber'];
      this.startDate = _data['startDate']
        ? new Date(_data['startDate'].toString())
        : <any>undefined;
    }
  }

  static fromJS(data: any): CreateCardCommand {
    data = typeof data === 'object' ? data : {};
    let result = new CreateCardCommand();
    result.init(data);
    return result;
  }

  toJSON(data?: any) {
    data = typeof data === 'object' ? data : {};
    data['bankCardId'] = this.bankCardId;
    data['name'] = this.name;
    data['cardNumber'] = this.cardNumber;
    data['startDate'] = this.startDate
      ? this.startDate.toISOString()
      : <any>undefined;
    return data;
  }
}

export interface ICreateCardCommand {
  bankCardId?: string | undefined;
  name?: string | undefined;
  cardNumber?: string | undefined;
  startDate?: Date;
}